Intermediate work on Header.
