+++ styles/fonts.css
@@ -31,7 +31,7 @@
 }
 
 @font-face {
-  font-family: "Font Awesome\ 6 Brands";
+  font-family: "Font Awesome 6 Brands";
   font-style: normal;
   font-weight: 400;
   font-display: block;
@@ -39,7 +39,7 @@
 }
 
 @font-face {
-  font-family: "Font Awesome\ 6 Free";
+  font-family: "Font Awesome 6 Free";
   font-style: normal;
   font-weight: 900;
   font-display: block;
