+++ styles/styles.css
@@ -91,6 +91,7 @@
   --font-family-heading: var(--font-family-body);
   --font-family-footer: var(--font-family-body);
   --font-family-fixed: 'Roboto Mono', menlo, consolas, 'Liberation Mono', monospace;
+  --font-family-fa-free: 'Font Awesome 6 Free';
 
   /* font weights */
   --lato-regular: 400;
@@ -120,7 +121,7 @@
   --icon-size-xxl: 100px;
 
   /* misc */
-  --nav-height: 5rem;
+  --nav-height: 78px;
 
   /* break points */
   --breakpoint-s: 576px; /* <= mobile */
@@ -162,14 +163,13 @@
     --heading-font-size-l: 26px;
     --heading-font-size-m: 20px;
     --section-width: var(--content-width-l);
-    --nav-height: 9.375rem;
   }
 }
 
 @media (min-width: 1200px /* breakpoint-xl */) {
   :root {
     --section-width: var(--content-width-xl);
-    --nav-height: 8.125rem;
+    --nav-height: 165px;
   }
 }
 
@@ -198,7 +198,6 @@ body.appear {
 
 header {
   height: var(--nav-height);
-  visibility: hidden;
 }
 
 h1, h2, h3,
